+++ src/widgets/dialog-provider/dialog-provider.stories.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import { DialogProvider, DialogProviderContext } from "./dialog-provider";
-import { DialogSize, DialogHelper } from "../../../types";
+import { DialogHelper, DialogSize } from "../../../types";
 import { Meta } from "@storybook/react";
 
 export default {
@@ -30,7 +30,9 @@ export const Default = (): any => {
     return (
       <div className="card">
         <div className="card-body">
-          <p>Click on a button to send the value to the component which launched it</p>
+          <p>
+            Click on a button to close the dialog and <strong>return a value to the host</strong>
+          </p>
 
           <button className="btn btn-secondary btn-sm" onClick={() => helper.dismiss("A")}>
             A
@@ -78,6 +80,26 @@ export const Default = (): any => {
     );
   };
 
+  const MessageDialog = ({ helper }: { helper: DialogHelper }): any => {
+    return (
+      <div className="card">
+        <div className="card-body">
+          <p>
+            Click the button to send a message to the host <strong>without closing the dialog</strong>
+          </p>
+
+          <button className="btn btn-primary btn-sm me-2" onClick={() => helper.send("Hello")}>
+            Send message (Hello)
+          </button>
+
+          <button className="btn btn-link btn-sm" onClick={helper.dismiss}>
+            Close dialog
+          </button>
+        </div>
+      </div>
+    );
+  };
+
   const Showcase = (): any => {
     const { showDialog } = useContext(DialogProviderContext);
 
@@ -106,42 +128,53 @@ export const Default = (): any => {
     };
 
     const openNonEscapeDismissibleDialog = (): void => {
-      openBackgroundDismissibleDialog();
       showDialog(helper => <NonEscapeDismissibleDialog helper={helper} />, { escapeDismissible: false });
+    };
+
+    const openMessageDialog = (): void => {
+      showDialog(helper => <MessageDialog helper={helper} />, {
+        onMessage: message => {
+          alert(`Dialog sent a message: ${message}`);
+        }
+      });
     };
 
     return (
       <div>
-        <button className="btn btn-primary btn-sm" onClick={(): void => openDialog(DialogSize.SMALL)}>
+        <button className="btn btn-primary btn-sm mb-2" onClick={(): void => openDialog(DialogSize.SMALL)}>
           Open small dialog
         </button>
         <span>&nbsp;</span>
-        <button className="btn btn-primary btn-sm" onClick={(): void => openDialog(DialogSize.MEDIUM)}>
+        <button className="btn btn-primary btn-sm mb-2" onClick={(): void => openDialog(DialogSize.MEDIUM)}>
           Open medium dialog
         </button>
         <span>&nbsp;</span>
-        <button className="btn btn-primary btn-sm" onClick={(): void => openDialog(DialogSize.WIDE)}>
+        <button className="btn btn-primary btn-sm mb-2" onClick={(): void => openDialog(DialogSize.WIDE)}>
           Open wide dialog
         </button>
         <span>&nbsp;</span>
-        <button className="btn btn-primary btn-sm" onClick={(): void => openDialog(DialogSize.FULL)}>
+        <button className="btn btn-primary btn-sm mb-2" onClick={(): void => openDialog(DialogSize.FULL)}>
           Open full dialog
         </button>
         <span>&nbsp;</span>
-        <button className="btn btn-info btn-sm" onClick={openAllDialogs}>
+        <button className="btn btn-info btn-sm mb-2" onClick={openAllDialogs}>
           Open all dialogs
         </button>
         <span>&nbsp;</span>
-        <button className="btn btn-success btn-sm" onClick={openDialogForValue}>
+        <button className="btn btn-success btn-sm mb-2" onClick={openDialogForValue}>
           Open dialog for value
         </button>
         <span>&nbsp;</span>
-        <button className="btn btn-warning btn-sm" onClick={openBackgroundDismissibleDialog}>
+        <button className="btn btn-warning btn-sm mb-2" onClick={openBackgroundDismissibleDialog}>
           Open background dismissible dialog
         </button>
         <span>&nbsp;</span>
-        <button className="btn btn-secondary btn-sm" onClick={openNonEscapeDismissibleDialog}>
+        <button className="btn btn-secondary btn-sm mb-2" onClick={openNonEscapeDismissibleDialog}>
           Open non-escape dismissible dialog
+        </button>
+        <span>&nbsp;</span>
+        <button className="btn btn-secondary btn-sm mb-2" onClick={openMessageDialog}>
+          Open message dialog
         </button>
       </div>
     );